refactor(spells): split prerequisites.test.ts along describe clusters (#522)
prerequisites.test.ts had grown to 716 lines (above the 500-line ceiling).
Split into four focused files matching the underlying module's concerns,
with shared mock factories extracted to helpers/prereq-fixtures.ts:

  - prerequisites-registry.test.ts (153) — collectPrerequisites (step-command),
    checkPrerequisites, formatPrerequisiteErrors
  - prerequisites-spec.test.ts (172) — compilePrerequisiteSpec detectors +
    collectPrerequisites YAML walker
  - prerequisites-resolve.test.ts (145) — resolveUnmetPrerequisites (TTY /
    non-TTY / abort / empty / promptOnMissing)
  - prerequisites-integration.test.ts (211) — SpellCaster integration +
    OAP retrofit (#518) end-to-end
  - helpers/prereq-fixtures.ts (87) — makePrereq, makeCommand, simpleSpell,
    createMockCredentials, createMockMemory, withEnvVars

vitest.config.ts isolationTests: single prerequisites.test.ts entry replaced
with all four split paths (same dynamic-import load profile).

Test count preserved 1:1 at 38 cases (11 + 11 + 7 + 9). Each new file passes
alone under vitest.isolation.config.ts; the @moflo/spells Pass 1 stays at
1143 tests, Pass 2 isolation batch at 167 passed / 10 skipped.

Incidental cleanups during the split:
  - withEnvVars(overrides, fn) helper replaces two OAP save/delete/restore
    try/finally blocks, eliminating the forget-to-restore-undefined branch.
  - loadOap() hoisted into beforeAll so YAML is parsed once per suite, not
    per-it; findOapPath uses async fsPromises.access instead of accessSync
    at module-collection time.
  - ENV_KEY hoisted in prerequisites-spec.test.ts with a beforeEach delete,
    removing four inlined cleanup lines and a thrown-assertion env leak.

Closes #522
